I comandi accanto a un dispositivo escono dal modello del robot
Le regole dei comandi in più — quali entità possono esserlo, come si chiamano
senza ripetere il nome dell'apparecchio, cosa fa toccarne una, quali stanno
accanto a quel dispositivo — sono nate dentro il modello del robot, perché lì
è arrivata la domanda per prima (#306). Poi se le sono prese anche gli
elettrodomestici, e nel loro file c'era scritto: «sono nate lì perché lì è
arrivata la domanda per prima; sono le stesse».

Adesso la stessa domanda arriva dalle TV (#451), e la terza volta quel
commento smette di essere una nota e diventa un difetto: un lettore che
importa dal robot non si spiega a chi legge. Le regole sono in un modulo loro,
e le tre sezioni le vedono senza sapere niente l'una dell'altra.

Insieme allo spostamento, i nomi dicono quello che le funzioni fanno:
`comandiDelRobot` → `comandiDelDispositivo`, `comandoDelRobot` →
`comandoDelDispositivo`, `comandiSuggeriti` → `comandiVicini`. Erano già
importate con un alias da chi non è un robot, e un alias è un nome sbagliato
con un cerotto sopra.

Nessun comportamento cambia: le prove che c'erano passano come prima, e quella
che pretendeva l'import dal modello del robot adesso pretende il modulo nuovo.
