+++ custom_components/dashboardmodern/frontend/legacy/dashboard.html
@@ -62,6 +62,7 @@ html[data-theme="dark"] #cd-boot-overlay b{color:#e2e8f0}
 <link rel="modulepreload" href="../src/core/cerca-nel-config.js">
 <link rel="modulepreload" href="../src/core/chiavi-di-configurazione.js">
 <link rel="modulepreload" href="../src/core/climate-power.js">
+<link rel="modulepreload" href="../src/core/comandi-accanto.js">
 <link rel="modulepreload" href="../src/core/come-sta-la-casa.js">
 <link rel="modulepreload" href="../src/core/contenuto-delle-sezioni.js">
 <link rel="modulepreload" href="../src/core/cover-kind.js">
