+++ custom_components/dashboardmodern/frontend/src/core/robot-model.js
@@ -16,6 +16,12 @@
  * specie decide quale dialetto si usa, senza che chi disegna debba saperlo.
  */
 
+import {
+  COMANDI_MASSIMI,
+  comandiDelDispositivo,
+  elencoComandi,
+  genereDelComando,
+} from "./comandi-accanto.js";
 import { conservaIlConfigurato } from "./device-model.js";
 import { SOURCE_LOCALE, getLocale, pick } from "./i18n.js";
 import { nomeAccantoAlDispositivo } from "./nome-accanto-al-dispositivo.js";
@@ -179,66 +185,6 @@ export function normalizeRobot(input = {}, index = 0) {
   );
 }
 
-/* ── i comandi a parte del robot (#306) ──────────────────────────────────
- *
- * «Le varie entita' del robot continuano a non essere visibili: da solo la
- * modalita' aspirazione. Comandi mancanti: button.roborock_..._asp_e_lav,
- * ..._pulizia_completa, ..._solo_aspirazione, ..._solo_lavaggio.»
- *
- * Un robot che lava non e' solo un `vacuum`. L'integrazione pubblica accanto a
- * lui i suoi programmi come tasti, le sue regolazioni — il mocio, l'acqua —
- * come tendine, le sue funzioni come interruttori: sono entita' a parte, e la
- * scheda del robot non le vedeva perche' guardava solo la sua. Qui si dice
- * quali entita' possono stare sulla scheda come comandi, come si chiamano
- * senza ripetere il nome del robot, e cosa fa ognuna quando la si tocca. Chi
- * configura le sceglie; chi disegna le mostra nell'ordine in cui sono scritte.
- */
-export const DOMINI_COMANDO = Object.freeze({
-  button: "tasto",
-  input_button: "tasto",
-  script: "tasto",
-  scene: "tasto",
-  /* Un robot comandato a automazioni e' un robot come gli altri.
-   *
-   * Non tutti i robot pubblicano dei «button»: chi ne ha uno che Home
-   * Assistant non integra a fondo si scrive le automazioni — Pulizia, Pausa,
-   * Dock, Pulizia programmata — ed e' quello il suo cruscotto. Lasciarle
-   * fuori voleva dire far nascere quel robot dall'integrazione con la riga
-   * dei comandi vuota. */
-  automation: "tasto",
-  switch: "interruttore",
-  input_boolean: "interruttore",
-  select: "tendina",
-  input_select: "tendina",
-});
-
-/* Dodici: una scheda e' una scheda, non la pagina delle impostazioni. */
-export const COMANDI_MASSIMI = 12;
-
-/** Che genere di comando e' quell'entita': tasto, tendina, interruttore — o niente. */
-export function genereDelComando(entity) {
-  return DOMINI_COMANDO[clean(entity).split(".")[0]] || "";
-}
-
-/** L'elenco pulito: solo entita' comandabili, una volta sola, non piu' di dodici. */
-export function elencoComandi(input) {
-  const grezzi = Array.isArray(input)
-    ? input
-    : typeof input === "string"
-      ? input.split(/[\s,;]+/)
-      : [];
-  const visti = new Set();
-  const fuori = [];
-  for (const voce of grezzi) {
-    const entity = clean(voce && typeof voce === "object" ? voce.entity : voce);
-    if (!entity || !genereDelComando(entity) || visti.has(entity)) continue;
-    visti.add(entity);
-    fuori.push(entity);
-    if (fuori.length >= COMANDI_MASSIMI) break;
-  }
-  return fuori;
-}
-
 export const ESITI_ELENCO = Object.freeze({
   aggiunto: "aggiunto",
   gia: "gia",
@@ -294,94 +240,6 @@ export function conLaVoce(elenco, nuovo, tipo = "comandi") {
   if (attuali.includes(entity)) return { elenco: attuali, esito: ESITI_ELENCO.gia };
   if (attuali.length >= regola.tetto) return { elenco: attuali, esito: ESITI_ELENCO.pieno };
   return { elenco: [...attuali, entity], esito: ESITI_ELENCO.aggiunto };
-}
-
-/** I comandi di un robot come stanno adesso: nome, genere, stato, opzioni. */
-export function comandiDelRobot(robot = {}, states = {}) {
-  return elencoComandi(robot?.comandi).map((entity) => {
-    const genere = genereDelComando(entity);
-    const corrente = states?.[entity];
-    const attributi = corrente?.attributes || {};
-    const stato = clean(corrente?.state).toLowerCase();
-    return {
-      entity,
-      genere,
-      name: nomeAccantoAlDispositivo(entity, robot, states),
-      /* Un tasto mai premuto sta su «unknown», ed e' un tasto che funziona:
-       * non raggiungibile e' solo chi lo dice. */
-      available: Boolean(corrente) && stato !== "unavailable",
-      acceso: genere === "interruttore" ? stato === "on" : null,
-      opzioni:
-        genere === "tendina" && Array.isArray(attributi.options)
-          ? attributi.options.map(clean).filter(Boolean)
-          : [],
-      scelta: genere === "tendina" ? clean(corrente?.state) : "",
-    };
-  });
-}
-
-/**
- * Il servizio dietro un comando.
- *
- * Un tasto si preme, uno script e una scena si accendono, un'automazione si fa
- * partire, un interruttore si inverte, una tendina sceglie: cinque verbi per
- * nove domini, e nessun servizio inventato — sono quelli che Home Assistant ha
- * per quelle entita'.
- */
-export function comandoDelRobot(voce = {}, valore = "") {
-  const entity = clean(voce?.entity);
-  const dominio = entity.split(".")[0];
-  const genere = genereDelComando(entity);
-  if (!genere) return null;
-  if (genere === "tendina") {
-    const opzione = clean(valore);
-    if (!opzione) return null;
-    return {
-      domain: dominio,
-      service: "select_option",
-      data: { entity_id: entity, option: opzione },
-    };
-  }
-  if (genere === "interruttore")
-    return { domain: dominio, service: "toggle", data: { entity_id: entity } };
-  if (dominio === "button" || dominio === "input_button")
-    return { domain: dominio, service: "press", data: { entity_id: entity } };
-  /* Un'automazione si fa PARTIRE, non si accende: «automation.turn_on» la
-   * riabilita e basta — il robot non si muove, e chi tocca il tasto ha appena
-   * cambiato di nascosto un'impostazione di Home Assistant. Il verbo giusto e'
-   * «trigger», ed e' l'unico che fa quello che il tasto promette. */
-  if (dominio === "automation")
-    return { domain: dominio, service: "trigger", data: { entity_id: entity } };
-  return { domain: dominio, service: "turn_on", data: { entity_id: entity } };
-}
-
-/* I comandi che stanno accanto al robot, da proporre a chi configura.
- *
- * Le entita' di uno stesso dispositivo si riconoscono da come Home Assistant
- * le chiama: l'id comincia con l'id del robot — `vacuum.roborock_qrevo` e
- * `button.roborock_qrevo_asp_e_lav` — oppure il nome comincia col nome del
- * robot. Sono proposte, non scelte: un robot pubblica anche i tasti che
- * azzerano i contatori dei filtri, e nessuno li vuole sotto il pollice senza
- * averlo detto. Chi configura li vede in scheda e tocca quelli che vuole. */
-export function comandiSuggeriti(robot = {}, states = {}) {
-  const entity = clean(robot?.entity);
-  const radice = entity.split(".")[1] || "";
-  if (!radice) return [];
-  const nome = clean(states?.[entity]?.attributes?.friendly_name).toLowerCase();
-  const gia = new Set(elencoComandi(robot?.comandi));
-  const trovati = [];
-  for (const [id, corrente] of Object.entries(states || {})) {
-    if (gia.has(id) || !genereDelComando(id)) continue;
-    const oggetto = id.split(".")[1] || "";
-    const stessoId = oggetto.startsWith(`${radice}_`);
-    const suoNome = clean(corrente?.attributes?.friendly_name).toLowerCase();
-    const stessoNome = Boolean(nome) && suoNome.startsWith(`${nome} `);
-    if (stessoId || stessoNome) trovati.push(id);
-  }
-  const ordine = { tasto: 0, tendina: 1, interruttore: 2 };
-  return trovati.sort(
-    (a, b) => ordine[genereDelComando(a)] - ordine[genereDelComando(b)] || a.localeCompare(b),
-  );
 }
 
 /* Le mappe che stanno accanto al robot, da proporre a chi configura (#468).
@@ -520,6 +378,12 @@ export const drawableRobots = (robots = []) =>
  * telecamera (e' cosi' che fanno Valetudo e le integrazioni Xiaomi) o come
  * un'immagine. In tutti e due i casi l'indirizzo del disegno sta in
  * `entity_picture`, e cambia a ogni aggiornamento. */
+export function robotMapPicture(states = {}, mapEntity = "") {
+  const reference = clean(mapEntity);
+  if (!reference) return "";
+  return clean(states?.[reference]?.attributes?.entity_picture);
+}
+
 /** Le mappe di un robot, ognuna col suo nome e il suo disegno di adesso. */
 export function mappeDelRobot(robot = {}, states = {}) {
   return elencoMappe(robot?.mappe ?? robot?.mapEntity).map((entity) => ({
@@ -527,12 +391,6 @@ export function mappeDelRobot(robot = {}, states = {}) {
     name: nomeAccantoAlDispositivo(entity, robot, states),
     picture: robotMapPicture(states, entity),
   }));
-}
-
-export function robotMapPicture(states = {}, mapEntity = "") {
-  const reference = clean(mapEntity);
-  if (!reference) return "";
-  return clean(states?.[reference]?.attributes?.entity_picture);
 }
 
 /**
@@ -590,7 +448,7 @@ export function robotView(robot = {}, states = {}) {
      * configura ha scelto, gia' scritte come vanno lette. */
     letture: lettureDelRobot(robot, states),
     /* I comandi a parte (#306), come stanno adesso. */
-    comandi: comandiDelRobot(robot, states),
+    comandi: comandiDelDispositivo(robot, states),
   };
 }
